Remove floating-point mod, and rename integer mod to remainder.
Floating-point mod is not in IEEE-754, not implemented in any hardware,
and not in C or C++ or similar languages. It's in asm.js, but probably
only because it's in JS, and it's in JS possibly only because JS Numbers
are intended to be somewhat usable as integers. Floating-point mod does
have use cases, however they can adequately be serviced by the math
library, in whatever form that ends up taking.

This patch also renames integer mod to remainder, since "signed modulus"
in particular is a non-obvious way to describe the mathematic operation
it performs.