Recategorize several integer operations "under consideration".
ctz/clz/popcount/bswap - not nicely polyfillable (ES6 will have clz, but
not all browsers will have that in the timeframe of the polyfill), but quite
essential in a modern ISA. These can be done in software and
pattern-matched, but except maybe ctz, the patterns are sufficiently gross
that I don't expect these will be controversial.

rot/min/max - nice to have, but not quite as essential, so put them in
FutureFeatures.md.