Recategorize the floating-point operations "under consideration".
min/max - Promote these to base; they are all trivially polyfillable,
and are useful things to have in the base platform besides.

nearestInt - This is not trivially polyfillable (JavaScript's Math.round
does tie-breaking wrong), so put it in EssentialPostMVPFeatures.md so it
can join floor/ceil.

trunc - This is not trivially polyfillable (in ES6, but not in all the
browsers we need the polyfill to support), so put it in
EssentialPostMVPFeatures.md so it can join floor/ceil/nearestInt in
providing rouding to integer via all the standard roundings.

minNum/maxNum - Put these in FutureFeatures.md for future consideration.
They're not trivially polyfillable, and they're not obviously essential
at this point, though they do have uses.