Describe integer operation trapping with a new rule.
The rule is that explicitly signed or unsigned operations trap any time
they can't return the actual result, and signed-less operations never
trap. There's no behavior change here because this covers exactly the
same cases that we already had covered.

The beauty of this rule is that it elides trapping for add, sub, and mul
for now, which would add too much overhead on today's hardware, while
leaving it in place for div and rem, where it isn't as problematic and
where we already had it, while simultaneously leaving a nice opening for
the possible future addition of explicitly signed and unsigned versions
of add, sub, and mul which would trap on overflow (and overflow is
signedness-dependent anyway).

This requires describing the right shifts as signedless; fortunately
there's already a common naming we can use, "logical" and "arithmetic".