Flesh out calls, indirect calls, and function pointers.
The function pointer type and its operations (conversions, comparisons)
were previously only implied. This patch documents them like regular
types and operations.

It also lays the groundwork for a future with multiple-return-value
functions. This is done by eliminating the `void` type and defining
return types as a sequence. In the MVP, the sequence can have at most
one value.