Added else child to the if statement.
Without an else branch we would need a temp local for writing normal if/else statements. If someone just wants an normal if statement he/she can just use a empty block as an else.

Other suggestion if changing `if` is not up to debate: Add an `ifelse` and keep `if` the way it is.