Make the sign bit of a new NaN value nondeterministic.
x86 produces NaN values with the sign bit set to 1, and other architectures
produce NaN values with the sign bit set to 0; wasm can model this by
just making the sign bit nondeterminisitc. This should still permit the
NaN-boxing technique that all this NaN bit-pattern language is intended to
support.