Make br_if return the value of its value operand.
This follows naturally from br_if's value operand being strict. We
evaluate the expression to produce a value even when the condition is
false, so it'd be nice to be able to use the value when the condition
is false too.

This change also makes sense from a low-level compiler perspective. The
value operand of a br_if represents a value "live out" of the br_if's
basic block. Since br_if is a simple conditional branch, the value
should be "live in" in both succesors.