Note that leading zeros in leb128 values have no semantic effect.
While this is implict in the current usage of the values, this clause
is intended to guard against future specification additions that
disguish wasm files based on redundant leading zeros. Also to guard
again consumers treating files differently based on such
choices. Perhaps also to require canonicalization if hashing for
resource integrity.