Use three ascii character long sections names for wasm sections.
This would allow the runtime to read an int32 and test this alone to
dispatch on the sections IDs that a runtime needs to read to execute
the wasm. It would include both the length and the three ascii letters
of the name. Obviously this strategy only works if the leb128 encoding
excludes redundant zeros, but it would only be slightly more complex
to read a variable length leb128 and ignore all sections without a
length of three and then dispatch on the three bytes of the name.

Also reserve short names for use in the wasm specification. This could
be relaxed in future but is intended to prevent a land-grab of short
names that the specification might want to use in future.