Use a fixed with encoding for sections sizes.
The section sizes will often need back-patching to avoid firtly
writing to a separate store to determine the size and then copying
across. For example, for the functions sections even if the AST sizes
are known when encoding it would take another pass to determine the
encoded size.

There are only a small number of sections so the file size difference
compared to using the leb128 encoding is expected to be small, and
this goes some way to avoiding the need for leb128 with redundant
leading zeros.