Move the function bodies OOL wrt the sections.
Only the function bodies remain inline in the sections as potential
bulk data. With the function bodies moved OOL it might be practical
for the runtime to delay compilation until all the section
definitions have been read and then to use the section defintions to
plan the compilation of the module.

The body sizes have been kept in the section definitions as a
suggestion. These would be expected to be back patched and perhaps
retain a fixed width encoding. Knowing the function body sizes early
might give the runtime compiler a good hint about the resources need
to compile each function. For example, it might choose to compile huge
functions with only a single compilation thread running if the runtime
is memory constrained.

The local variable definitions have been moved OOL as a
suggestion. These might not be known when building the sections and
having them OOL might avoid back patching their definition into the
sections which would become more problematic with the new layout for
these.