Remove `loop`'s bottom label.
This simplifies `loop` by removing its bottom label, so that it
only has a top label. This establishes a nice consistency in
having every scoped construct -- `block`, `loop`, `if`, `else`,
and function body -- add exactly one "level" to the scope depth.

When a `loop` exit label is needed, an explicit `block` can be
used. In practice, `loop` is typically less than 1% of all nodes,
and beyond that, the majority of `loop`s in practice don't use
their bottom label anyway.

In the s-expression format with both labels at the top, I often
forget which label is for the top and which is for the bottom. And
it can be awkward in other possible formats as well, requiring
either multiple labels or additional disambiguation.

And while it's not complicated to implement loop as a two-level
construct (having done it myself, both as a producer and as a
consumer), it is an odd special case.