Make dropping of values explicit (#694)
Simplify Wasm and allow a more natural interpretation of Wasm as a stack machine. The main changes are:

- Values can no longer be discarded implicitly.
- Instead, there is an explicit drop operator.
- To compensate, store operators no longer return a value.

The constructs affected by this are mainly blocks and block-like sequences. Before, all expressions in a block could yield a value, and it would just be dropped on the floor implicitly (except the last one). Now we require explicit drop operations in those cases. With stores no longer returning a value, the only places were we expect drops to actually arise are calls to side-effecting functions that also return a value, but this value isn't used.

(Also fixing a bunch of other out-of-date text on the way.)