WebAssembly.compile: neuter received buffer
It doesn't seem useful for the buffer to be available for any purpose after it's been handed off to the `compile` function. The current wording forces the implementation to copy the entire buffer, which is silly. Neutering the provided buffer allows for a more efficient implementation.

My suggested wording is similar to the wording for Worker.postMessage.