JS API algorithms: use numbering (#871)
It's easier to refer to JS APIs using numbered lists rather than bullet indentations. This isn't perfect because Markdown tops out after some amount of numbering, but it's better than before!

I think we'll separately want to rejigger some of the ordering of the algorithms. This change should be a no-op.