Explain validation
Validation is currently only defined in the spec, which is not approachable by everybody. This PR describes the basic idea.

Based on various online and offline discussions, it also makes the handling of unreachable code more precise and addresses a couple of concerns that keep coming up with the status quo:

1. it allows implementations to diverge to some degree, by not requiring them to fully check,
2. it is not easy to understand for everybody what that lenience implies,
3. non-lenient validators need a certain amount of inference, especially with potential future feature like `dup`.

The description in this PR makes precise the treatment of code after branches and tries to adopt the simplest possible rules. For most current implementations, it should not imply any real change. The spec interpreter and the test suite would need some adjustment (removing the "soft failure" mode).