Export name uniqueness
spec test exports.wast makes it a validation failure for exported names to be duplicate. I find this somewhat surprising (I see exports as a dictionary, so duplicates simply overwrite previous entries), I'd therefore be happy if others want to reverse the polarity of this (make it OK to duplicate export names).