WebAssembly.instantiate could be specified more explicitly (#1018)
* WebAssembly.instantiate could be specified more explicitly

The current wording for WebAssembly.instantiate is a little confusing. It's not clear exactly what it means to run a task that does JS asynchronously as JS is not specified to be asynchronous. This change clarifies that instantiate should queue a task to do the JS work.

Additionally, an implementor may wish to do more asynchronous work after they have looked at a module's imports. This change also explicitly allows implementors to do more asynchronous work.