Add Typed Object support to WebAssembly
For interoperability between normal JavaScript and WebAssembly, Typed Objects
can be used as a bridge because the object's memory layout is defined by its
typed descriptor. Typed objects are always unboxed and therefore provide fast
read and write access from both WebAssembly and normal JavaScript. The "core"
design of Typed Object is mandatory for implementing object support in
WebAssembly, while the object oriented features are not.

Support for interoperability between arbitrary JavaScript objects and
WebAssembly is not feasible at the moment. Typed Object provide a way of
interfacing with the same memory without worrying about whether the value is
the real value or stored in a NaN-boxed format.