Delete unnecessary sections
As garbage collection, multiple return, and Memset/Memcpy Operators have
dedicated tracking issues and all the links have been fixed, we can
remove their sections. Most of the data being removed is still present
in the tracking issues.

The two sections on tail calls are not removed as they contain some
information not in the tail calls tracking issue (#1144).