Delete outdated pre-standard langauge description text.
Delete all the old content in Semantics.md, TextFormat.md, JS.md,
BinaryFormat.md, and Modules.md, so that they just contain links to
the relevant chapters of the official spec.

The old content was out of date, in some places no longer accurate and
in others no longer relevant, so having it in the tree is confusing
for people who are searching for things and possibly missing the notices
about being unmaintained at the tops of the files.

All the content is still available in the git history in case anyone
wants to do archeology.