Update README.md, Nondeterminism.md, and Security.md
Rewrite README.md to reflect the design repo's current purposes and
update outdated information. Also, remove text from README.md which
now lives in the webassembly.org website, and link to the website
instead.

Also, update some outdated information and links in Nondeterminism
and Security.md, so that we can link to them as examples of fairly
current documents.